rename ShareFile to FileService and update package structure; remove DatabaseSetup class and add DatabaseUtil and FileUtil classes
