--- Util/DatabaseSetup.java
@@ -1,5 +1,0 @@
-package Util;
-
-public class DatabaseSetup {
-    
-}
